errors

+++ controllers/cards.js
@@ -1,37 +1,69 @@
 const Card = require('../models/card');
 
-const getCards = (req, res, next) => {
+const getCards = (req, res) => {
   Card.find({})
     .then((cards) => res.status(200).send(cards))
-    .catch(next);
+    .catch((e) => res.status(500).send({ message: `Ошибка получения карточек ${e}` }));
 };
 const deleteCard = (req, res, next) => {
   const { cardId } = req.params;
   Card.findById(cardId)
-    .then(() => {
-      Card.findByIdAndRemove(cardId)
-        .then((deletedCard) => res.status(200).send(deletedCard))
-        .catch(next);
+    .then((card) => {
+      if (!card) {
+        res.status(404).send({ message: 'Карточка не найдена' });
+      } else {
+        Card.findByIdAndRemove(cardId)
+          .then((deletedCard) => res.status(200).send(deletedCard))
+          .catch(next);
+      }
     })
-    .catch(next);
+    .catch((e) => res.status(500).send({ message: `Ошибка удаления карточки ${e}` }));
 };
 const createCards = (req, res) => {
   const { name, link } = req.body;
-  Card.create({ name, link, owner: req.user._id }).then((card) => res.status(200).send(card));
+  Card.create({ name, link, owner: req.user._id })
+    .then((card) => res.status(201).send(card))
+    .catch((e) => {
+      if (e.name === 'ValidationError') {
+        res.status(400).send({ message: 'Переданы неверные данные' });
+      } else {
+        res.status(500).send({ message: `Ошибка создания карточки ${e}` });
+      }
+    });
 };
 
 const likeCard = (req, res) => {
-  Card.findByIdAndUpdate(
-    req.params.cardId,
-    { $addToSet: { likes: req.user._id } },
-    { new: true },
-  ).then((card) => res.status(200).send(card));
+  Card.findByIdAndUpdate(req.params.cardId, { $addToSet: { likes: req.user._id } }, { new: true })
+    .then((card) => {
+      if (!card) {
+        res.status(404).send({ message: 'Карточка не найдена' });
+      }
+      res.status(200).send(card);
+    })
+    .catch((e) => {
+      if (e.name === 'CastError') {
+        res.status(400).send({ message: 'Переданы неверные данные' });
+      } else {
+        res.status(500).send({ message: `Ошибка ${e}` });
+      }
+    });
 };
 
 const dislikeCard = (req, res) => {
-  Card.findByIdAndUpdate(req.params.cardId, { $pull: { likes: req.user._id } }, { new: true }).then(
-    (card) => res.status(200).send(card),
-  );
+  Card.findByIdAndUpdate(req.params.cardId, { $pull: { likes: req.user._id } }, { new: true })
+    .then((card) => {
+      if (!card) {
+        res.status(404).send({ message: 'Карточка не найдена' });
+      }
+      res.status(200).send(card);
+    })
+    .catch((e) => {
+      if (e.name === 'CastError') {
+        res.status(400).send({ message: 'Переданы неверные данные' });
+      } else {
+        res.status(500).send({ message: `Ошибка ${e}` });
+      }
+    });
 };
 module.exports = {
   getCards,
